feat: 매니저용 매칭 요청 검토 페이지 추가 - mockData로 UI 확인 가능, 추후 API 연동 예정
refactor: 라우팅 구조 및 중복 파일 정리
- reservationId, managerId 등 동적 세그먼트 라우트 충돌 방지 위해 manager/, reservation/ 하위로 이동
- [reservationId]/result, [managerId] 등 중복/불필요 파일 삭제
- 공통 컴포넌트는 components/로 통합
