+++ apps/src/app/(standalone)/matching/reservation/[reservationId]/page.tsx
@@ -1,3 +1,7 @@
+// 매칭 상세 페이지: 선택한 매니저의 상세 정보, 예약 정보, 메시지, 추가 옵션을 확인하고 매칭을 수락/거절할 수 있는 페이지
+// 매니저의 매칭 수락 후 사용자가 최종 수락/거절을 선택하는 페이지 -> 모달로 구현 
+// 삭제해도 될 거 같은데 확인 필요
+
 'use client';
 
 import Image from 'next/image';
