+++ apps/src/app/(standalone)/matching/reservation/[reservationId]/result/loading.tsx
@@ -1,0 +1,5 @@
+// 매칭 결과 페이지의 로딩 상태를 표시하는 스켈레톤 UI 컴포넌트
+
+import MatchingResultLoading from '../../../components/loading';
+
+export default MatchingResultLoading;
