--- apps/src/app/matching/[reservationId]/result/loading.tsx
@@ -1,3 +1,0 @@
-import MatchingResultLoading from '../../components/loading';
-
-export default MatchingResultLoading;
